pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hkstock-quant"
version = "0.1.0"
description = "高可用、可扩展的港股量化交易系统"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "pandas>=1.5.3",
    "numpy>=1.24.3",
    "scipy>=1.10.1",
    "pytz>=2023.3",
    "backtesting>=0.3.3",
    "backtrader>=1.9.76.123",
    "yfinance>=0.2.18",
    "akshare>=1.10.0",
    "pandas-datareader>=0.10.0",
    "requests>=2.29.0",
    "matplotlib>=3.7.1",
    "seaborn>=0.12.2",
    "plotly>=5.14.1",
    "dash>=2.9.3",
    "dash-bootstrap-components>=1.4.1",
    "futu-api>=7.1.3300",
    "alpaca-py>=0.8.2",
    "scikit-learn>=1.2.2",
    "statsmodels>=0.14.0",
    "tensorflow>=2.12.0",
    "keras>=2.12.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "pytest>=7.3.1",
    "flask>=2.3.2",
    "flask-restful>=0.3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "jupyter>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
